+++ phidgets_magnetometer/src/magnetometer_ros_i.cpp
@@ -81,8 +81,8 @@ MagnetometerRosI::MagnetometerRosI(const rclcpp::NodeOptions& options)
             "work");
     }
 
-    publish_rate_ = this->declare_parameter("publish_rate", 0);
-    if (publish_rate_ > 1000)
+    publish_rate_ = this->declare_parameter("publish_rate", 0.0);
+    if (publish_rate_ > 1000.0)
     {
         throw std::runtime_error("Publish rate must be <= 1000");
     }
@@ -179,9 +179,9 @@ MagnetometerRosI::MagnetometerRosI(const rclcpp::NodeOptions& options)
     magnetometer_pub_ =
         this->create_publisher<sensor_msgs::msg::MagneticField>("imu/mag", 1);
 
-    if (publish_rate_ > 0)
+    if (publish_rate_ > 0.0)
     {
-        double pub_msec = 1000.0 / static_cast<double>(publish_rate_);
+        double pub_msec = 1000.0 / publish_rate_;
         timer_ = this->create_wall_timer(
             std::chrono::milliseconds(static_cast<int64_t>(pub_msec)),
             std::bind(&MagnetometerRosI::timerCallback, this));
@@ -334,7 +334,7 @@ void MagnetometerRosI::magnetometerChangeCallback(
         last_data_timestamp_ns_ = this_ts_ns;
 
         // Publish if we aren't publishing on a timer
-        if (publish_rate_ <= 0)
+        if (publish_rate_ <= 0.0)
         {
             publishLatest();
         }
